+++ vagrant/test/fixtures/cookbooks/test/recipes/windows_vagrant_plugin.rb
@@ -1,0 +1,3 @@
+include_recipe 'vagrant'
+
+vagrant_plugin 'vagrant-ohai'